requirements.txt
fastapi>=0.110
uvicorn>=0.24
streamlit>=1.31
requests>=2.31
opencv-python>=4.8
numpy>=1.24
pillow>=10.0
torch>=2.1
torchvision>=0.16
ftfy>=6.1
regex>=2023.10.3
tqdm>=4.66
faiss-cpu>=1.7.4
scenedetect[opencv]>=0.6
git+https://github.com/openai/CLIP.git
